chore: improve e2e teardown debugging, change hasura data button scraping
